+++ src/probeinterface/__init__.py
@@ -14,9 +14,6 @@ from .io import (
     write_csv,
     read_BIDS_probe,
     write_BIDS_probe,
-    read_spikegadgets,
-    read_spikegadgets_neuropixels,
-    has_spikegadgets_neuropixels_probes,
     read_mearec,
     read_nwb,
     read_maxwell,
@@ -33,6 +30,10 @@ from .neuropixels_tools import (
     read_openephys_neuropixels,
     has_neuropixels_probes,
     get_saved_channel_indices_from_openephys_settings,
+    read_spikegadgets,
+    read_spikegadgets_neuropixels,
+    has_spikegadgets_neuropixels_probes,
+    parse_spikegadgets_header,
 )
 from .utils import combine_probes
 from .generator import (
